fix: 일부 API 조정
