+++ src/sdk/crack/session-fetcher-util.ts
